Codegen/Rust: allow breaking up schema file
