+++ rust/codegen.conf
@@ -1,4 +1,5 @@
 # configuration file for Rust code generation default options
+--schema=schema
 --generate=dbscheme,rusttest,ql,rust
 --dbscheme=ql/lib/rust.dbscheme
 --ql-output=ql/lib/codeql/rust/elements/internal/generated
