+++ rust/schema/__init__.py
@@ -1,0 +1,16 @@
+"""
+Schema description
+
+This file should be kept simple:
+* no flow control
+* no aliases
+* only class definitions with annotations and `include` calls
+
+For how documentation of generated QL code works, please read `misc/codegen/schema_documentation.md`.
+"""
+
+from .prelude import *
+from .ast import *
+
+include("../shared/tree-sitter-extractor/src/generator/prefix.dbscheme")
+include("prefix.dbscheme")
