+++ rust/schema/ast.py
@@ -1,60 +1,4 @@
-"""
-Schema description
-
-This file should be kept simple:
-* no flow control
-* no aliases
-* only class definitions with annotations and `include` calls
-
-For how documentation of generated QL code works, please read `misc/codegen/schema_documentation.md`.
-"""
-
-from misc.codegen.lib.schemadefs import *
-
-include("../shared/tree-sitter-extractor/src/generator/prefix.dbscheme")
-include("prefix.dbscheme")
-
-
-@qltest.skip
-class Element:
-    pass
-
-
-@qltest.skip
-class Locatable(Element):
-    pass
-
-
-@qltest.skip
-class AstNode(Locatable):
-    pass
-
-
-@qltest.skip
-class Unextracted(Element):
-    """
-    The base class marking everything that was not properly extracted for some reason, such as:
-    * syntax errors
-    * insufficient context information
-    * yet unimplemented parts of the extractor
-    """
-    pass
-
-
-@qltest.skip
-class Missing(Unextracted):
-    """
-    The base class marking errors during parsing or resolution.
-    """
-
-
-@qltest.skip
-class Unimplemented(Unextracted):
-    """
-    The base class for unimplemented nodes. This is used to mark nodes that are not yet extracted.
-    """
-    pass
-
+from .prelude import *
 
 class AssocItem(AstNode):
     pass
